Modules reorganized, targo-bench now is not dependent on criterion
