+++ examples/benches/criterion.rs
