+++ examples/benches/tango-faster.rs
@@ -1,8 +1,11 @@
 #![cfg_attr(feature = "align", feature(fn_align))]
 
 use crate::test_funcs::{factorial, sum};
-use tango_bench::{benchmark_fn, tango_benchmarks, tango_main, BenchmarkMatrix, IntoBenchmarks};
-use test_funcs::{sort_unstable, str_count, str_take, RandomSubstring, RandomVec};
+use tango_bench::{
+    benchmark_fn, generators::RandomVec, tango_benchmarks, tango_main, BenchmarkMatrix,
+    IntoBenchmarks,
+};
+use test_funcs::{sort_unstable, str_count, str_take, RandomSubstring};
 
 mod test_funcs;
