+++ tango-bench/Cargo.toml
@@ -25,26 +25,11 @@ goblin = "^0.7"
 scroll = "^0.11"
 tempfile = "^3.8"
 
-[dev-dependencies]
-criterion = "^0.5"
-
 [features]
 align = []
 hw_timer = []
 aa_test = []
 
 [[bench]]
-name = "criterion"
-harness = false
-
-[[bench]]
 name = "tango"
-harness = false
-
-[[bench]]
-name = "tango-faster"
-harness = false
-
-[[bench]]
-name = "tango-slower"
 harness = false
